LayoutTests/imported/w3c/web-platform-tests/inert/inert-with-modal-dialog-001-expected.txt


PASS Modal dialog only marks outside nodes as inert
PASS Inner nodes with 'inert' attribute become inert anyways
PASS If the modal dialog has the 'inert' attribute, everything becomes inert
FAIL If an ancestor of the dialog has the 'inert' attribute, everything becomes inert assert_equals: expected "" but got "dialog child"
wrapper
